Provide no defaults for buildDescriptorFile param
The problem here is that if no condition is activated, the default of the first declaration is used, i.e. "pom.xml", even though it may be wrong, i.e. for buildTool=mta. The versioning package already has the same defaults built-in, so it is not necessary to fill the parameter.